requirements.txt
opencv-python
numpy
matplotlib
seaborn
pandas
jsonschema
torch
torchvision
facenet-pytorch
scikit-learn
pillow
